// src/interfaces/product/UserPrivateProductEstimates.ts
interface UserPrivateProductEstimates {
  profileId: number;
  productId: number;
  storyMark?: number;
  artMark?: number;
  infoMark?: number;
  visualMark?: number;
  soundMark?: number;
  actMark?: number;
  gameplayMark?: number;
  spentTime?: number;
  rating: number;
}

export type { UserPrivateProductEstimates };